Move rubygems aggregate source to a separate class
It allows to use a consistent source interface in the resolver.
